requirements.txt
google-cloud-firestore
speedtest-cli
ruamel.yaml
pythonping
telethon
aiohttp
aiofiles
click
black
